+++ base-components/videoPlayer/types.ts
